+++ src/settings/FeeFinesTable/index.js
@@ -1,2 +1,3 @@
 export { default as Owners } from './Owners';
 export { default as CopyModal } from './CopyModal';
+export { default as ChargeNotice } from './ChargeNotice';