+++ CHANGELOG.md
@@ -19,6 +19,9 @@
 * Fix labels on Pay and Waive modals. Fix UIU-845.
 * Update circulation v7.0 and request-storage v3.0 OKAPI interfaces. Part of UIU-889.
 * Use packageInfo passed via props if present. Part of ERM-72.
+* Add feature "Notify patron of new manual fee/fine charge". Fix UIU-710.
+* Add feature "Notify patron of fee/fine payment received". Fix UIU-711.
+* Add feature "Charge notice and action notice options to manual fee/fine settings screen". Fix UIU-713.
 
 ## [2.20.0](https://github.com/folio-org/ui-users/tree/v2.20.0) (2019-01-25)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.19.0...v2.20.0)